chore: index.js sugar
